removed unnecessary interface
