--- core/src/com/jingyuyao/tactical/model/state/BaseControllingState.java
@@ -1,44 +1,0 @@
-package com.jingyuyao.tactical.model.state;
-
-import com.google.common.base.Preconditions;
-import com.jingyuyao.tactical.model.ModelBus;
-import com.jingyuyao.tactical.model.event.Save;
-import com.jingyuyao.tactical.model.ship.Ship;
-import com.jingyuyao.tactical.model.world.Cell;
-
-class BaseControllingState extends BaseState implements ControllingState {
-
-  private final StateFactory stateFactory;
-  private final Cell cell;
-  private final Ship ship;
-
-  BaseControllingState(
-      ModelBus modelBus, WorldState worldState, StateFactory stateFactory, Cell cell) {
-    super(modelBus, worldState);
-    Preconditions.checkArgument(cell.ship().isPresent());
-    Preconditions.checkArgument(cell.ship().get().isControllable());
-    this.stateFactory = stateFactory;
-    this.cell = cell;
-    this.ship = cell.ship().get();
-  }
-
-  @Override
-  public Cell getCell() {
-    return cell;
-  }
-
-  @Override
-  public Ship getShip() {
-    return ship;
-  }
-
-  StateFactory getStateFactory() {
-    return stateFactory;
-  }
-
-  void finish() {
-    ship.setControllable(false);
-    post(new Save());
-    branchTo(stateFactory.createWaiting());
-  }
-}
