+++ core/src/com/jingyuyao/tactical/model/state/ControllingState.java
@@ -1,17 +1,48 @@
 package com.jingyuyao.tactical.model.state;
 
+import com.google.common.base.Preconditions;
+import com.jingyuyao.tactical.model.ModelBus;
+import com.jingyuyao.tactical.model.event.Save;
 import com.jingyuyao.tactical.model.ship.Ship;
 import com.jingyuyao.tactical.model.world.Cell;
 
-public interface ControllingState extends State {
+public class ControllingState extends BaseState {
+
+  private final StateFactory stateFactory;
+  private final Cell cell;
+  private final Ship ship;
+
+  ControllingState(
+      ModelBus modelBus, WorldState worldState, StateFactory stateFactory, Cell cell) {
+    super(modelBus, worldState);
+    Preconditions.checkArgument(cell.ship().isPresent());
+    Preconditions.checkArgument(cell.ship().get().isControllable());
+    this.stateFactory = stateFactory;
+    this.cell = cell;
+    this.ship = cell.ship().get();
+  }
 
   /**
    * The {@link Cell} being controlled.
    */
-  Cell getCell();
+  public Cell getCell() {
+    return cell;
+  }
 
   /**
    * The {@link Ship} being controlled.
    */
-  Ship getShip();
+  public Ship getShip() {
+    return ship;
+  }
+
+  StateFactory getStateFactory() {
+    return stateFactory;
+  }
+
+  void finish() {
+    ship.setControllable(false);
+    post(new Save());
+    branchTo(stateFactory.createWaiting());
+  }
 }
